# apps/interlabs-crm-demo/backend/openapi.yaml
openapi: 3.0.3
info:
  title: Interlab Portal API
  version: 1.0.0
  description: |
    Internal Enterprise Dashboard — CRM + ERP + Realtime Operations Hub for
    PT. Interlab Sentra Solutions Indonesia.

    Covers: Auth, Users & Roles, PO lifecycle (11 stages across 5 divisions),
    Sales, Admin & Log, Finance, Technical, HRGA/Legal, Tax & Insurance,
    Notifications, Chat, Email Templates, Files.

    ## Response envelope
    Every response uses a consistent envelope:
      success: `{ "success": true,  "data": ..., "meta"?: {...} }`
      error:   `{ "success": false, "error": "message", "code"?: "..." }`

    ## RBAC
    Every protected endpoint declares `x-rbac` with the feature_key,
    capability, allowed role_keys, and row-level scope. The RBAC middleware
    enforces `(feature, capability)` against `role_permissions` on every
    request; `scope` is applied to queries via the sameRoleScope /
    row-level-security layer.

    ## PO lifecycle side-effects
    Endpoints marked `x-po-side-effects` MUST, in a single transaction:
      1. Append to `purchase_order_status_history`
      2. Append to `purchase_order_tracking_events`
      3. Update `purchase_orders.current_status`
      4. Emit the declared notification templates (if enabled)
  contact:
    name: Interlab Portal Engineering
    email: zakyindrasatriaputra@gmail.com
servers:
  - url: https://api.interlab-portal.com
    description: Production
  - url: http://localhost:3000
...
security:
  - bearerAuth: []
paths:
...
  /api/tax/operational:
    get:
      tags: [Tax & Insurance]
      summary: List tax operational records
      x-rbac: { feature: tax_operational, capability: view_own, allowed_roles: [superadmin, ceo, tax_insurance], scope: own }
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/SearchParam'
        - { in: query, name: tax_type, schema: { type: string, enum: [PPh 21, PPh 25, PPN, Others] } }
        - { in: query, name: tax_category, schema: { type: string, enum: [SSP Payment, SPT Reporting, Combined Record] } }
        - { in: query, name: masa_pajak_month, schema: { type: integer, minimum: 1, maximum: 12 } }
        - { in: query, name: masa_pajak_year, schema: { type: integer } }
        - { in: query, name: tahun_pajak, schema: { type: integer } }
        - { in: query, name: record_status, schema: { type: string, enum: [Draft, Submitted, Verified, Archived] } }
        - { in: query, name: payment_status, schema: { type: string, enum: [Unpaid, Paid, Pending, Failed] } }
        - { in: query, name: pic_user_id, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Paginated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data:
                        type: array
                        items: { $ref: '#/components/schemas/TaxOperationalRecord' }
    post:
      tags: [Tax & Insurance]
      summary: Create tax operational record
      description: Appends to tax_operational_audit_log with action=created.
      x-rbac: { feature: tax_operational, capability: create, allowed_roles: [superadmin, ceo, tax_insurance], scope: own }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaxOperationalCreateRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/TaxOperationalRecord' }
  /api/tax/operational/{id}:
    parameters: [ { $ref: '#/components/parameters/UUIDPathId' } ]
    get:
      tags: [Tax & Insurance]
      summary: Get tax operational record
      x-rbac: { feature: tax_operational, capability: view_own, allowed_roles: [superadmin, ceo, tax_insurance], scope: own }
      responses:
        '200':
          description: Record
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/TaxOperationalRecord' }
    put:
      tags: [Tax & Insurance]
      summary: Update tax operational record
      description: Appends to tax_operational_audit_log with action=updated or status_changed.
      x-rbac: { feature: tax_operational, capability: edit, allowed_roles: [superadmin, ceo, tax_insurance], scope: own }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TaxOperationalCreateRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessEnvelope'
                  - properties:
                      data: { $ref: '#/components/schemas/TaxOperationalRecord' }
    delete:
      tags: [Tax & Insurance]
      summary: Soft-delete tax operational record
      x-rbac: { feature: tax_operational, capability: delete, allowed_roles: [superadmin, ceo], scope: global }
      responses:
        '204': { description: Deleted }

  # ====================================================================
  # Notifications
  # ====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (1h expiry). Obtained from `POST /api/auth/login`.
        Refresh with `POST /api/auth/refresh` (refresh token 7d, or 30d with remember_me).
  parameters:
    PageParam:
      in: query
      name: page
      required: false
      schema: { type: integer, minimum: 1, default: 1 }
    LimitParam:
      in: query
      name: limit
      required: false
      schema: { type: integer, enum: [25, 50, 75, 100], default: 25 }
    SearchParam:
      in: query
      name: search
      required: false
      schema: { type: string }
...
    UUIDPathId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [success, data]
      properties:
        success: { type: boolean, enum: [true] }
        data: {}
        meta: { $ref: '#/components/schemas/PaginationMeta' }
...
    PaginationMeta:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        total_pages: { type: integer }

    # ------------------------------------------------------------------
    # Enums
    # ------------------------------------------------------------------
...
    Currency:
      type: string
      enum: [IDR, USD, EUR]
...
    AuditFields:
      type: object
      properties:
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
        created_by: { type: string, format: uuid, nullable: true, readOnly: true }
        updated_by: { type: string, format: uuid, nullable: true, readOnly: true }
        deleted_at: { type: string, format: date-time, nullable: true, readOnly: true }
...
    TaxOperationalRecord:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
        - type: object
          properties:
            id: { type: string, format: uuid, readOnly: true }
            tax_operational_record_number: { type: string, readOnly: true }
            tax_type: { type: string, enum: [PPh 21, PPh 25, PPN, Others] }
            tax_category: { type: string, enum: [SSP Payment, SPT Reporting, Combined Record] }
            masa_pajak: { type: string, format: date, nullable: true, description: First day of the reporting month }
            masa_pajak_month: { type: integer, minimum: 1, maximum: 12, nullable: true }
            masa_pajak_year: { type: integer, nullable: true }
            tahun_pajak: { type: integer, nullable: true }
            npwp: { type: string, description: 15-digit Indonesian tax ID }
            taxpayer_name: { type: string, nullable: true }
            taxpayer_address: { type: string, nullable: true }
            jenis_spt: { type: string, enum: [SPT Tahunan, SPT Masa], nullable: true }
            status_spt: { type: string, enum: [Normal, Pembetulan], nullable: true }
            reporting_date: { type: string, format: date, nullable: true }
            billing_code: { type: string, nullable: true }
            ntpn: { type: string, nullable: true }
            ntb: { type: string, nullable: true }
            stan: { type: string, nullable: true }
            bank_name: { type: string, nullable: true }
            payment_date: { type: string, format: date, nullable: true }
            amount: { type: number, nullable: true }
            currency: { $ref: '#/components/schemas/Currency' }
            payment_status: { type: string, enum: [Unpaid, Paid, Pending, Failed] }
            record_status: { type: string, enum: [Draft, Submitted, Verified, Archived] }
            pic_user_id: { type: string, format: uuid, nullable: true }
            notes: { type: string, nullable: true }
            attachment_ssp_file_ids:
              type: array
              items: { type: string, format: uuid }
            attachment_spt_file_ids:
              type: array
              items: { type: string, format: uuid }
            attachment_payment_file_ids:
              type: array
              items: { type: string, format: uuid }
            attachment_supporting_file_ids:
              type: array
              items: { type: string, format: uuid }
    TaxOperationalCreateRequest:
      type: object
      required: [tax_type, tax_category, npwp]
      properties:
        tax_type: { type: string, enum: [PPh 21, PPh 25, PPN, Others] }
        tax_category: { type: string, enum: [SSP Payment, SPT Reporting, Combined Record] }
        masa_pajak: { type: string, format: date }
        tahun_pajak: { type: integer }
        npwp: { type: string }
        taxpayer_name: { type: string }
        taxpayer_address: { type: string }
        jenis_spt: { type: string, enum: [SPT Tahunan, SPT Masa] }
        status_spt: { type: string, enum: [Normal, Pembetulan] }
        reporting_date: { type: string, format: date }
        billing_code: { type: string }
        ntpn: { type: string }
        ntb: { type: string }
        stan: { type: string }
        bank_name: { type: string }
        payment_date: { type: string, format: date }
        amount: { type: number }
        currency: { $ref: '#/components/schemas/Currency' }
        payment_status: { type: string, enum: [Unpaid, Paid, Pending, Failed] }
        record_status: { type: string, enum: [Draft, Submitted, Verified, Archived] }
        pic_user_id: { type: string, format: uuid }
        notes: { type: string }
        attachment_ssp_file_ids:
          type: array
          items: { type: string, format: uuid }
        attachment_spt_file_ids:
          type: array
          items: { type: string, format: uuid }
        attachment_payment_file_ids:
          type: array
          items: { type: string, format: uuid }
        attachment_supporting_file_ids:
          type: array
          items: { type: string, format: uuid }

    # ------------------------------------------------------------------
    # Notifications
    # ------------------------------------------------------------------
...